+++ core/src/main/java/com/shatteredpixel/yasd/items/quest/Pickaxe.java
@@ -64,12 +64,12 @@ public class Pickaxe extends Weapon {
 	public boolean bloodStained = false;
 
 	@Override
-	public int min(int lvl) {
+	public int min(float lvl) {
 		return 2;   //tier 2
 	}
 
 	@Override
-	public int max(int lvl) {
+	public int max(float lvl) {
 		return 15;  //tier 2
 	}
