+++ core/src/main/java/com/shatteredpixel/yasd/items/weapon/melee/RunicBlade.java
@@ -35,7 +35,7 @@ public class RunicBlade extends MeleeWeapon {
 	//equal to tier 4 in damage at +5
 
 	@Override
-	public int max(int lvl) {
+	public int max(float lvl) {
 		return  (int) (5*(tier) +                	//20 base, down from 25
 				Math.round(lvl*(tier*2.5)));	//+6 per level, up from +5
 	}
